+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/point.rb
@@ -2,7 +2,11 @@ module ActiveRecord
   module ConnectionAdapters
     module PostgreSQL
       module OID # :nodoc:
-        class Point < Type::String
+        class Point < Type::Value
+          def type
+            :point
+          end
+
           def type_cast(value)
             if ::String === value
               if value[0] == '(' && value[-1] == ')'