+++ activerecord/test/cases/adapters/postgresql/geometric_test.rb
@@ -15,9 +15,9 @@ class PostgresqlPointTest < ActiveRecord::TestCase
     @connection = ActiveRecord::Base.connection
     @connection.transaction do
       @connection.create_table('postgresql_points') do |t|
-        t.column :x, :point
-        t.column :y, :point, default: [12.2, 13.3]
-        t.column :z, :point, default: "(14.4,15.5)"
+        t.point :x
+        t.point :y, default: [12.2, 13.3]
+        t.point :z, default: "(14.4,15.5)"
       end
     end
   end
@@ -28,9 +28,9 @@ class PostgresqlPointTest < ActiveRecord::TestCase
 
   def test_column
     column = PostgresqlPoint.columns_hash["x"]
-    assert_equal :string, column.type
+    assert_equal :point, column.type
     assert_equal "point", column.sql_type
-    assert column.text?
+    assert_not column.text?
     assert_not column.number?
     assert_not column.binary?
     assert_not column.array
@@ -48,9 +48,9 @@ class PostgresqlPointTest < ActiveRecord::TestCase
 
   def test_schema_dumping
     output = dump_table_schema("postgresql_points")
-    assert_match %r{t\.string\s+"x"$}, output
-    assert_match %r{t\.string\s+"y",\s+default: \[12\.2, 13\.3\]$}, output
-    assert_match %r{t\.string\s+"z",\s+default: \[14\.4, 15\.5\]$}, output
+    assert_match %r{t\.point\s+"x"$}, output
+    assert_match %r{t\.point\s+"y",\s+default: \[12\.2, 13\.3\]$}, output
+    assert_match %r{t\.point\s+"z",\s+default: \[14\.4, 15\.5\]$}, output
   end
 
   def test_roundtrip